+++ META.json
@@ -70,6 +70,7 @@
          "requires" : {
             "Attribute::Handlers" : "0",
             "Config" : "0",
+            "Data::Dump" : "0",
             "File::Spec" : "0",
             "Sub::Util" : "0",
             "XSLoader" : "0",
